engine: extract esm/{healpix_tile,ui_tools}.js
